requirements.txt
crewai==0.36.1
Django==4.2.11
djangorestframework==3.15.2
langchain_community==0.2.7
langchain_groq==0.1.6
